+++ package.json
@@ -31,6 +31,7 @@
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "express-graphql": "^0.9.0",
+    "express-jwt": "^5.3.1",
     "graphql": "^14.5.8",
     "jsonwebtoken": "^8.5.1",
     "mongodb": "^3.4.1",
